velome_chat/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
python-dotenv
pydantic
sentence-transformers
tiktoken
groq
